Move exception from common
